fix: 修复所有14个Kotlin编译错误
- AppUseCases: 添加PackageInfo import
- FileUseCases: Long转Int
- GameBoosterUseCases: 添加suspend修饰符
- NetworkUseCases: 重构async并发逻辑
- CryptoUseCases: 添加pow import, override name
- DeepInspectorUseCases: 添加BatteryManager/IntentFilter import

+++ app/src/main/java/com/banana/toolbox/domain/usecase/file/FileUseCases.kt
@@ -118,7 +118,7 @@ class FileUseCases @Inject constructor(
                 totalBytes = totalBytes,
                 usedBytes = usedBytes,
                 freeBytes = freeBytes,
-                usagePercent = if (totalBytes > 0) (usedBytes * 100 / totalBytes) else 0
+                usagePercent = if (totalBytes > 0) (usedBytes * 100 / totalBytes).toInt() else 0
             )
         }
     }
